+++ modules/camera/include/o3d3xx_camera/device_config.h
@@ -188,6 +188,16 @@ namespace o3d3xx
     bool EnableAcquisitionFinishedPCIC() const noexcept;
     void SetEnableAcquisitionFinishedPCIC(bool on) noexcept;
 
+    int EIPProducingSize() const noexcept;
+    void SetEIPProducingSize(int i) noexcept;
+
+    int EIPConsumingSize() const noexcept;
+    void SetEIPConsumingSize(int i) noexcept;
+
+    bool PcicTcpSchemaAutoUpdate() const noexcept;
+    void SetPcicTcpSchemaAutoUpdate(bool on) noexcept;
+
+
   protected:
     /**
      * User-defined name of the device (max 64 characters)
@@ -351,24 +361,42 @@ namespace o3d3xx
     double temp_illu_;
 
     /**
-     * @todo needs documentation
+     * The Profinet device name 
      */
     std::string pnio_device_name_;
 
     /**
-     * @todo needs documentation
+     * Select the ethernet fieldbus 
      */
     int ethernet_field_bus_;
 
     /**
-     * @todo needs documentation
+     * Selects the Endianess of the fieldbus
      */
     int ethernet_field_bus_endianness_;
 
     /**
-     * @todo needs documentation
+     * Enable the asynchronous notification for the
+     * acquisition finished signal over PCIC
      */
     bool enable_acquisition_finished_pcic_;
+
+    /**
+     * Ethernet/IP producing assembly size
+     */
+    int eip_producing_size_;
+
+    /**
+     * Ethernet/IP consuming assembly size
+     */
+    int eip_consuming_size_;
+
+    /**
+     * This enables the PCIC auto schema update.
+     * When enabled the default schema will be updated on all
+     * open connections
+     */
+    bool pcic_tcp_schema_auto_update_;
 
   }; // end: class DeviceConfig